+++ CHANGELOG.md
@@ -22,6 +22,7 @@ And list your changes under that.
 - Fixes a bug where interventions were too persistent and ignored the user's preferences in one sec.
 - Adds a brand new search page that lets you search for users, statuses, and tags.
 - Introduces a new Focused Timeline setting to split timeline lists between "original" content and replies/reblogs.
+- Updates composer window activation to be more consistent with the Mail app.
 
 ## [1.0-14 (beta)] - 15/1/2023